==> Initialization/02. Housing/00. Init Secret Census Key.ipynb <==

# Init Secret Census Key

This script is designed to securely store your Census Bureau API Key, which is required for accessing many of the data services provided by the U.S. Census Bureau.

To obtain an API key from the U.S. Census Bureau, you will need to apply at the following website: [Census Bureau API Key Signup](https://api.census.gov/data/key_signup.html).

After obtaining your API key, run this script and enter your key when prompted. The script will save your key to a pickle file named `secrets.pkl`. This file should be stored securely and not shared or made publicly accessible.

To ensure the file is not accidentally committed to a public GitHub repository, add the filename to your `.gitignore` file. Open the `.gitignore` file in a text editor, add a new line with `secrets.pkl`, and save the file. If your repository does not have a `.gitignore` file, you can create one.


import getpass
import pickle

# Prompt the user for their API key. The input will be hidden.
api_key = getpass.getpass("Please enter your census API key:")

# Save the key to a pickle file
with open('secrets.pkl', 'wb') as f:
    pickle.dump(api_key, f)

print("API key stored successfully.")
